Go sources rearrange
